Changed ExternalBookingId from Long to Integer and adjusted the Flatly Dto's.

+++ backend/src/main/java/pw/react/backend/dto/parkly/ParklyBookingDetailsResponse.java
@@ -11,7 +11,7 @@ import java.util.List;
 public class ParklyBookingDetailsResponse {
 
     private Integer bookingId;
-    private Long externalBookingId;
+    private Integer externalBookingId;
     private String status;
 
     private LocalDateTime dateFrom;
