+++ backend/src/main/java/pw/react/backend/services/parkly/ParklyMainService.java
@@ -111,7 +111,7 @@ public class ParklyMainService implements ParklyService {
                 });
     }
     @Transactional
-    public boolean cancelCarBooking(Long externalBookingId) {
+    public boolean cancelCarBooking(Integer externalBookingId) {
         User parklyUser = userRepository.findByEmail("parkly@pw.edu.pl")
                 .orElseThrow(() -> new ResourceNotFoundException("Parkly system user not found. Seed data missing."));
 
@@ -135,7 +135,7 @@ public class ParklyMainService implements ParklyService {
     }
     @Override
     @Transactional(readOnly = true)
-    public ParklyBookingDetailsResponse getCarBookingByExternalBookingId(Long externalBookingId) {
+    public ParklyBookingDetailsResponse getCarBookingByExternalBookingId(Integer externalBookingId) {
 
         User parklyUser = userRepository.findByEmail(PARKLY_SYSTEM_EMAIL)
                 .orElseThrow(() -> new ResourceNotFoundException("Parkly system user not found. Seed data missing."));
